python-saju/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
korean-lunar-calendar==0.3.1
mysql-connector-python==8.1.0
python-dotenv==1.0.0
